LogicMBTI

+++ src/main/java/com/example/mbti/repository/QuestionRepository.java
@@ -1,9 +1,9 @@
 package com.example.mbti.repository;
+
+
+import com.example.mbti.model.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.mbti.model.Question; 
 
-
-
-public interface QuestionRepository  extends JpaRepository<Question, Long> {
-   
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
 }
+
